requirements.txt
aiogram>=3.4.1,<4.0.0
discord.py>=2.3.2,<3.0.0
